Refactor user-ID to use Integer

+++ backend/src/main/java/com/activekids/web/services/UserService.java
@@ -6,5 +6,6 @@ import com.activekids.web.model.User;
 public interface UserService {
     Iterable<User> getAllUsers();
     boolean createUser(User user);
-    boolean removeUser(Long id);
+    boolean removeUser(Integer id);
+
 }
